Poprawki w formularzach

+++ frontend/src/components/cell/EditCellForm.js
@@ -8,10 +8,10 @@ import config from '../../config.json';
 
 export default function EditCellForm(props){
 
-    const [limit, setLimit] = useState(0);
+    const [limit, setLimit] = useState(props.cell ? props.cell.limit : undefined);
 
     const [limitList, setLimitList] = useState([]);
-
+    
     useEffect(() => {
         axios.get(config.API_URL + 'limit').then(response => {
 			setLimitList(response.data);
@@ -19,31 +19,27 @@ export default function EditCellForm(props){
     }, []);
 
     function save(){
-        if(limit < 0){
+        if(!Number.isInteger(limit) || limit < 0){
             NotificationManager.error('Niepoprawny limit', 'Błąd');
             return;
         }
         const cellList = limitList.filter(l => l.column === props.column.id && l.row === props.row.id);
-        if(cellList.length === 0){
-            if(limit > 0){
-                axios.post(config.API_URL + 'limit/', {
-                    column: props.column.id,
-                    row: props.row.id,
-                    limit: limit
-                }).then(response => {
-                    NotificationManager.success('Limit dodany', 'Powiadomienie');
-                    props.onClose();
-                    props.onSave();
-                })
-                .catch(error => {
-                    NotificationManager.error('Limit niedodany', 'Błąd');
-                });
-                return;
-            }
-            props.onClose();
+        if(!props.cell && limit > 0){
+            axios.post(config.API_URL + 'limit/', {
+                column: props.column.id,
+                row: props.row.id,
+                limit: limit
+            }).then(response => {
+                NotificationManager.success('Limit dodany', 'Powiadomienie');
+                props.onClose();
+                props.onSave();
+            })
+            .catch(error => {
+                NotificationManager.error('Limit niedodany', 'Błąd');
+            });
             return;
         }
-        if(limit === 0){
+        if(props.cell && limit === 0){
             axios.delete(config.API_URL + 'limit/' + cellList[0].id + '/').then(response => {
                 NotificationManager.success('Limit usunięty', 'Powiadomienie');
                 props.onClose();
@@ -54,16 +50,20 @@ export default function EditCellForm(props){
             });
             return;
         }
-        axios.patch(config.API_URL + 'limit/' + cellList[0].id + '/', {
-            limit: limit
-        }).then(response => {
-            NotificationManager.success('Limit zapisany', 'Powiadomienie');
-            props.onClose();
-            props.onSave();
-        })
-        .catch(error => {
-            NotificationManager.error('Limit niezapisany', 'Błąd');
-        });
+        if(props.cell && limit > 0 && props.cell.limit !== limit){
+            axios.patch(config.API_URL + 'limit/' + cellList[0].id + '/', {
+                limit: limit
+            }).then(response => {
+                NotificationManager.success('Limit zapisany', 'Powiadomienie');
+                props.onClose();
+                props.onSave();
+            })
+            .catch(error => {
+                NotificationManager.error('Limit niezapisany', 'Błąd');
+            });
+            return;
+        }
+        props.onClose();
     }
 
     return(
@@ -77,7 +77,7 @@ export default function EditCellForm(props){
                 <Form>
                     <Form.Group className='mb-3'>
                         <Form.Label>Limit zadań na komórkę</Form.Label>
-                        <Form.Control type='number' placeholder='Wpisz limit' min='0' onChange={(e) => setLimit(parseInt(e.target.value))}/>
+                        <Form.Control type='number' placeholder='Wpisz limit' min='0' onChange={(e) => setLimit(parseInt(e.target.value))} defaultValue={props.cell ? props.cell.limit : undefined}/>
                     </Form.Group>
                 </Form>
             </Modal.Body>
